[tests] mongodb final changes

+++ packages/syft/tests/utils/mongodb.py
@@ -18,6 +18,7 @@ import subprocess
 from tarfile import TarFile
 from tempfile import gettempdir
 from tempfile import mkdtemp
+import zipfile
 
 # third party
 import distro
@@ -26,7 +27,7 @@ import requests
 
 MONGO_CONTAINER_PREFIX = "pytest_mongo"
 MONGO_VERSION = "7.0"
-MONGO_FULL_VERSION = f"{7.0}.6"
+MONGO_FULL_VERSION = f"{MONGO_VERSION}.6"
 PLATFORM_ARCH = platform.machine()
 PLATFORM_SYS = platform.system()
 DISTRO_MONIKER = distro.id() + distro.major_version() + distro.minor_version()
@@ -34,6 +35,7 @@ DISTRO_MONIKER = distro.id() + distro.major_version() + distro.minor_version()
 MONGO_BINARIES = {
     "Darwin": f"https://fastdl.mongodb.org/osx/mongodb-macos-{PLATFORM_ARCH}-{MONGO_FULL_VERSION}.tgz",
     "Linux": f"https://fastdl.mongodb.org/linux/mongodb-linux-{PLATFORM_ARCH}-{DISTRO_MONIKER}-{MONGO_FULL_VERSION}.tgz",
+    "Windows": f"https://fastdl.mongodb.org/windows/mongodb-windows-x86_64-{MONGO_FULL_VERSION}.zip",
 }
 
 
@@ -63,13 +65,14 @@ def stop_mongo_server(name):
 
 def __start_mongo_proc(name, port):
     prefix = f"mongo_{name}_"
-    download_dir = Path(mkdtemp(prefix=prefix))
-    db_path = Path(mkdtemp(prefix=prefix))
+
+    download_dir = Path(gettempdir(), "mongodb")
 
     exec_path = __download_mongo(download_dir)
     if not exec_path:
         raise Exception("Failed to download MongoDB binaries")
 
+    db_path = Path(mkdtemp(prefix=prefix))
     proc = subprocess.Popen(
         [
             str(exec_path),
@@ -95,17 +98,28 @@ def __download_mongo(download_dir):
     if url is None:
         raise NotImplementedError(f"Unsupported platform: {PLATFORM_SYS}")
 
-    download_path = Path(download_dir, "mongodb.tgz")
+    download_path = Path(download_dir, f"mongodb_{MONGO_FULL_VERSION}.archive")
+    download_path.parent.mkdir(parents=True, exist_ok=True)
 
-    with requests.get(url, stream=True) as r:
-        r.raise_for_status()
-        with open(download_path, "wb") as f:
-            copyfileobj(r.raw, f)
+    if not download_path.exists():
+        # download the archive
+        with requests.get(url, stream=True) as r:
+            r.raise_for_status()
+            with open(download_path, "wb") as f:
+                copyfileobj(r.raw, f)
 
-    with TarFile.open(download_path) as tf:
-        tf.extractall(download_dir)
+        # extract it
+        if url.endswith(".zip"):
+            archive = zipfile.ZipFile(download_path, "r")
+        else:
+            archive = TarFile.open(download_path, "r")
 
-    for path in download_dir.glob("**/mongod"):
+        archive.extractall(download_dir)
+        archive.close()
+
+    for path in download_dir.glob(f"**/*{MONGO_FULL_VERSION}*/bin/mongod*"):
+        if path.suffix not in (".exe", ""):
+            continue
         return path
